requirements.txt
llama-index-core==0.12.52.post1
llama-index-instrumentation==0.4.0
llama-index-llms-ollama==0.6.2
llama-index-workflows==1.2.0
python-dotenv==1.1.1
nest-asyncio==1.6.0
openai==0.27.10  
aiohttp==3.12.15
tiktoken==0.9.0
numpy==2.3.2
pydantic==2.11.7
